+++ src/Registrars/RegisterMacros.php
@@ -7,11 +7,17 @@ namespace Josemontano1996\LaravelOctaneLocalization\Registrars;
 use Illuminate\Support\Facades\Route;
 use Josemontano1996\LaravelOctaneLocalization\Contracts\LocalizationConfigInterface;
 use Josemontano1996\LaravelOctaneLocalization\Middlewares\LocalizationMiddleware;
+use Josemontano1996\LaravelOctaneLocalization\Middlewares\LocalizationMiddlewareWithoutRedirect;
 
 class RegisterMacros
 {
     public static function register(): void
     {
+        Route::macro(
+        'localizedWithoutPrefix',
+        fn(callable|string $callback) => Route::middleware(LocalizationMiddlewareWithoutRedirect::class)->group($callback)
+    );
+
         Route::macro('localizedWithPrefix', function ($callback = null) {
             $config = app(LocalizationConfigInterface::class);
             $key = $config->getParameterKey();